changes function names and resolved some bugs
