+++ app/data_preparation/Python/precalculate_grid_thematic.py
@@ -4,10 +4,15 @@ import math
 from variables_precalculate import *
 from pathlib import Path
 import yaml
-from random import randint
+
+#Step defined the bulk size of the heatmap calculation
+step = 150
+#Grid size defines the size of each grid
+grid_size = 500
+beta = -0.003
+grid = 'grid_'+str(500)
 
 start = time.time()
-
 with open(str(Path.home())+"/app/config/goat_config.yaml", 'r') as stream:
     config = yaml.load(stream)
 secrets = config["DATABASE"]
@@ -21,88 +26,91 @@ password = secrets["PASSWORD"]
 con = psycopg2.connect("dbname='%s' user='%s' port = '%s' host='%s' password='%s'" % (
     db_name, user, port, host, password))
 cursor = con.cursor()
-grid_size = 500
 
 
-def calculate_isochrones(grid_size):
-    grid_size = grid_size
-    snap_tolerance = math.sqrt(2)*grid_size
-    cursor.execute(prepare_tables.replace('grid_size', str(grid_size)))
+cursor.execute(prepare_tables.replace('grid_size', str(grid_size)))
+con.commit()
+
+sql_ordered_grid = '''DROP TABLE IF EXISTS grid_ordered;
+CREATE temp TABLE grid_ordered AS 
+SELECT starting_points, grid_id
+FROM (
+    SELECT ARRAY[ST_X(st_centroid(geom))::numeric,ST_Y(ST_Centroid(geom))::numeric] starting_points, grid_id 
+    FROM grid_500 
+    ORDER BY st_centroid(geom)
+) x;
+ALTER TABLE grid_ordered ADD COLUMN id serial;
+ALTER TABLE grid_ordered ADD PRIMARY key(id);'''
+
+cursor.execute(sql_ordered_grid)
+con.commit()
+
+cursor.execute('SELECT count(*) FROM grid_ordered;')
+count_grids = cursor.fetchall()[0][0]
+
+print('Bulk calculation is starting...')
+
+lower_limit = 1
+while lower_limit < count_grids:
+    print(lower_limit)
+ 
+    sql_bulk_calculation = '''
+        WITH x AS 
+    	(
+    		SELECT array_agg(starting_points) AS array_starting_points, array_agg(grid_id) AS grid_ids
+    		FROM grid_ordered 
+    		WHERE id BETWEEN %i AND %i
+    	)
+    		SELECT precalculate_grid('grid_500',15, x.array_starting_points, 5, x.grid_ids) 
+    		FROM x;'''
+    cursor.execute(sql_bulk_calculation % (lower_limit, lower_limit+step-1))
     con.commit()
-    cursor.execute(
-        'select grid_id, st_x(st_centroid(geom)), st_y(st_centroid(geom)) from grid_%s order by grid_id' % (grid_size))
-    grid = cursor.fetchall()
-    for i in grid:
-        grid_id = int(i[0])
-        x = i[1]
-        y = i[2]
-        objectid = randint(1, 10000000)
-        insert_into = '''
-        DROP TABLE IF EXISTS iso_precalculate;
-        CREATE temp TABLE iso_precalculate AS WITH iso AS ( 
-        SELECT *, %i
-        FROM isochrones(999999,15,%f,%f,1,%f,%f,1,%i,1)
-        )
-        ,
-        upt AS (
-        INSERT INTO isochrones(userid,id,step,geom,objectid) SELECT * FROM iso
-        )
-        SELECT * FROM iso;
-        UPDATE grid_%s SET isochrone_gid = i.gid FROM isochrones i 
-        WHERE i.objectid = %i	
-        AND grid_id = %i;''' % (objectid, x, y, 83.33, 1, objectid,str(grid_size), objectid, grid_id)
-  
-        sql_update_grid= '''UPDATE grid_%i set area_isochrone = st_area(i.geom::geography)
-               from isochrones i
-               where isochrone_gid=i.gid;
-               UPDATE grid_%i SET pois = i.pois::jsonb 
-               FROM isochrones i WHERE i.objectid = %i
-               AND grid_id = %i''' % (grid_size,grid_size,objectid,grid_id)
-        try:
-            cursor.execute(insert_into)
-            cursor.execute('SELECT isochrone_gid FROM grid_500 WHERE grid_id = %i' % grid_id)
-            isochrone_gid = cursor.fetchall()
-            isochrone_gid = isochrone_gid[0][0]
-            print(isochrone_gid)
-            cursor.execute('SELECT * FROM thematic_data_json(%i)' % (isochrone_gid))
-            
-            cursor.execute(sql_update_grid)
-            con.commit()
-        except Exception as e:
-        	print(e)
-        	cursor.execute('rollback;')
-        continue
+    lower_limit = lower_limit + step
 
 
-def calculate_index(grid_size):
-    grid_size = int(grid_size)
 
-    cursor.execute(
-        "select variable_array from variable_container where identifier = 'poi_categories'")
-    pois = cursor.fetchall()[0][0]
+print('Expanding jsonb to columns is starting...')
 
-    for sensitivity in sensitivities:
+sql_poi_categories = '''SELECT poi FROM (
+SELECT unnest(variable_array::text[]) poi FROM variable_container WHERE identifier = 'poi_categories' 
+) x 
+ORDER BY poi;
+'''
 
-        cursor.execute('ALTER TABLE grid_%i add column index_0_%s jsonb' % (grid_size, sensitivity))
-        con.commit()
-        cursor.execute('''UPDATE grid_%s set index_0_%s='{"a":"first_value"}'::jsonb''' % (grid_size, sensitivity))
-
-        sql_expand_json = 'select g.grid_id, '
-
-        for poi in pois:
-            cursor.execute(sql_calculate_accessibility % (float('0.'+sensitivity),grid_size,poi,poi,grid_size,sensitivity,sensitivity,grid_size))
-            con.commit()
-
-        cursor.execute('''UPDATE grid_%i set index_0_%s= index_0_%s - 'a' ''' % (grid_size,sensitivity,sensitivity))
-        con.commit()
+cursor.execute(sql_poi_categories)
+poi_categories = cursor.fetchall()
 
 
-calculate_isochrones(grid_size)
-calculate_index(grid_size)
+
+
+for i in poi_categories:
+    cursor.execute('ALTER TABLE %s ADD COLUMN %s numeric[];' % (grid,i[0]))
+    cursor.execute('''UPDATE %s set %s= jsonb_arr2text_arr((pois ->> '%s')::jsonb)::numeric[];''' % (grid,i[0],i[0]))
+
+print('Precalculate index is starting...')
+
+column_name_index = str(-beta).replace('.','_')
+cursor.execute('ALTER TABLE %s ADD COLUMN index_%s jsonb;' % (grid,column_name_index))
+cursor.execute('''UPDATE %s SET index_%s = '{}'::jsonb;''' % (grid,column_name_index))
+
+sql_compute_index = '''
+UPDATE %s SET index_%s = index_%s || jsonb_build_object('%s',x.accessibility_index)
+FROM (
+     SELECT grid_id, accessibility_index 
+    from heatmap_dynamic('[{"%s":{"sensitivity":%s,"weight":1}}]')
+    WHERE accessibility_index <> 0
+) x 
+WHERE %s.grid_id = x.grid_id
+'''
+
+for i in poi_categories:
+    cursor.execute(sql_compute_index % (grid,column_name_index,column_name_index,i[0],i[0],beta,grid))
 
 cursor.execute(sql_grid_population.replace('grid_size', str(grid_size)))
+
 con.commit()
 con.close()
 end = time.time()
 print('Running the script took:')
 print(end - start)
+
