+++ app/data_preparation/SQL/buildings_residential.sql
@@ -81,7 +81,6 @@ LEFT JOIN non_residential_ids n
 ON  x.osm_id = n.osm_id
 WHERE n.gid IS NULL; 
 
---All Building with no levels get building_levels = 2 AND roof_levels = 1
 UPDATE buildings_residential 
 set building_levels = (SELECT variable_simple::integer FROM variable_container WHERE identifier = 'default_building_levels'), 
 roof_levels = 1
