+++ app/data_preparation/SQL/create_tables.sql
@@ -152,7 +152,7 @@ values('poi_categories',
 "ice_cream","restaurant","theatre","sum_population","cinema","library","night_club","recycling",
 "car_sharing","bicycle_rental","charging_station","bus_station","tram_station","subway_station","railway_station","taxi",
 "hairdresser","atm","bank","dentist","doctors","pharmacy","post_box","post_office","fuel",
-"bakery","butcher","clothes","convenience","fashion","florist","greengrocer",
+"bakery","butcher","clothes","convenience","general","fashion","florist","greengrocer",
 "kiosk","mall","shoes","sports","supermarket","health_food","discount_supermarket",
 "hypermarket","international_supermarket","chemist","organic","marketplace",
 "hotel","museum","hostel","guest_house","viewpoint","gallery","bus_stop",
